Modelleme/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
jupyter
pyyaml
xgboost
niapy
lightgbm
joblib
openpyxl
